+++ lib/hanami/cli/banner.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "hanami/cli/program_name"
 
 module Hanami
@@ -81,8 +83,8 @@ module Hanami
         required_arguments = command.required_arguments
         optional_arguments = command.optional_arguments
 
-        required = required_arguments.map { |arg| arg.name.upcase }.join(' ') if required_arguments.any?
-        optional = optional_arguments.map { |arg| "[#{arg.name.upcase}]" }.join(' ') if optional_arguments.any?
+        required = required_arguments.map { |arg| arg.name.upcase }.join(" ") if required_arguments.any?
+        optional = optional_arguments.map { |arg| "[#{arg.name.upcase}]" }.join(" ") if optional_arguments.any?
         result = [required, optional].compact
 
         " #{result.join(' ')}" unless result.empty?