+++ spec/unit/hanami/cli/version_spec.rb
@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 RSpec.describe "Hanami::CLI::VERSION" do
   it "exposes version" do
-    expect(Hanami::CLI::VERSION).to eq("0.3.0")
+    expect(Hanami::CLI::VERSION).to eq("1.0.0.alpha1")
   end
 end