+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Hanami::CLI
 General purpose Command Line Interface (CLI) framework for Ruby
 
+## v1.0.0.alpha1 - 2019-01-30
+### Added
+- [Luca Guidi] Inheritng from subclasses of `Hanami::CLI::Command`, allows to inherit arguments, options, description, and examples.
+- [Luca Guidi] Allow to use `super` from `#call`
+
+### Changed
+- [Luca Guidi] Drop support for Ruby: MRI 2.3, and 2.4.
+
+## v0.3.1 - 2019-01-18
+### Added
+- [Luca Guidi] Official support for Ruby: MRI 2.6
+- [Luca Guidi] Support `bundler` 2.0+
+
 ## v0.3.0 - 2018-10-24
 
 ## v0.3.0.beta1 - 2018-08-08